image_api/requirements.txt
Flask
flask-cors
tensorflow
pillow
faiss-cpu
requests